requirements.txt
amqp==5.1.0
asgiref==3.5.0
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.2
Deprecated==1.2.13
Django==3.2
django-celery-results==2.3.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
flower==1.0.0
humanize==4.0.0
idna==3.3
importlib-metadata==4.11.3
iniconfig==1.1.1
kombu==5.2.4
model-mommy==2.0.0
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.1
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
py==1.11.0
pydevd-pycharm==221.4165.171
pyparsing==3.0.7
pytest==7.1.1
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
redis==4.1.4
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tornado==6.1
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.0
zipp==3.7.0
